Move UbuntuDesktopInstallerApp out of main.dart
To avoid having to import `main.dart` in various pages.
