+++ ubuntu_desktop_installer/lib/pages/try_or_install_page.dart
@@ -5,7 +5,7 @@ import 'package:flutter/rendering.dart';
 import 'package:flutter_html/flutter_html.dart';
 import 'package:url_launcher/url_launcher.dart';
 
-import '../main.dart';
+import '../app.dart';
 import '../routes.dart';
 import '../widgets.dart';
