+++ ubuntu_desktop_installer/test/widget_test.dart
@@ -10,7 +10,7 @@ import 'package:provider/provider.dart';
 
 import 'package:ubuntu_desktop_installer/keyboard_model.dart';
 import 'package:ubuntu_desktop_installer/l10n/app_localizations.dart';
-import 'package:ubuntu_desktop_installer/main.dart';
+import 'package:ubuntu_desktop_installer/app.dart';
 import 'package:ubuntu_desktop_installer/pages/welcome_page.dart';
 
 void main() {
